Rename key-storage.js --> credentials.js
